Move camera state migration to migration module
